1282 refactor persistence methods
