+++ shared/src/persistence/s3/zipDocuments.js
@@ -1,6 +1,5 @@
-const s3Zip = require('s3-zip');
+const s3Zip = require('./s3-zip');
 const stream = require('stream');
-
 /**
  * zipDocuments
  *
@@ -11,7 +10,14 @@ const stream = require('stream');
  * @param {string} providers.zipName the name of the generated zip file
  * @returns {Promise} the created zip
  */
-exports.zipDocuments = ({ applicationContext, fileNames, s3Ids, zipName }) => {
+exports.zipDocuments = ({
+  applicationContext,
+  extraFileNames,
+  extraFiles,
+  fileNames,
+  s3Ids,
+  zipName,
+}) => {
   return new Promise((resolve, reject) => {
     const { region } = applicationContext.environment;
     const bucket = applicationContext.environment.documentsBucketName;
@@ -44,6 +50,8 @@ exports.zipDocuments = ({ applicationContext, fileNames, s3Ids, zipName }) => {
         '',
         s3Ids,
         fileNames,
+        extraFiles,
+        extraFileNames,
       )
       .pipe(uploadFromStream(s3Client));
   });
